Refactoring. Handled "licensed video" error
