+++ ytsubtitles/main.py
@@ -1,3 +1,5 @@
 from ytsubtitles.subtitles_downloader import download_subs_arguments
 
-__all__ = ["download_subs_arguments"]
+
+if __name__ == '__main__':
+    download_subs_arguments()
